refactor: add init functions
